dev: apply comment suggestions
